Added functionality to gracefully finish at certain hour
Whenever the next file is being considered for an upload, if a user-defined hour is reached, the script will gracefully finish.  This is better than a hard cut-off (process kill). This has practical implications - you may not want to be using your organisational bandwidth for uploads for backups during working time. In future, it would also be good to implement bandwidth control.